fix(ui): links to elements by id shouldn't be not found
We switched from a browser router to a hash router so requests go
through the router in builds, which is basically a static file server
environment. However, a hash router uses the hash portion of the URL
to manage the routing. This means we can't use that to link to
elements by ID. The memory router seems to meet both of these
challenges.